requirements.txt
pandas==2.1.2
numpy>=1.25,<1.26
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
faker==19.13.0
pyyaml==6.0.1
python-dotenv==1.0.0
pytest==7.4.0
pytest-cov==4.1.0
loguru==0.7.2
